+++ R/utils.R
@@ -1,46 +1,36 @@
 # condition & helper functions # https://adv-r.hadley.nz/debugging
 
-# ceiling on decimal
-ceiling2 = function(x) {
-  x_sci = format(x, scientific = TRUE, digits=2)
-  z = ceiling(as.numeric(substr(x_sci, 1, 3)))
-  e = sub('.+(e.+)$', '\\1', x_sci)
-  as.numeric(paste0(z, e))
-}
-
-# is date/time class
-isdatetime = function(x) {
-  any(class(x) %in% c("Date","POSIXlt","POSIXct","POSIXt"))
-}
-
-
 # remove constant columns
+#' @import xefun
 check_const_cols = function(dt) {
   setDT(dt)
   # constant columns
-  const_cols = names(which(dt[,sapply(.SD, function(x) length(unique(x))==1)]))
+  const_cols = cols_const(dt)
   if (length(const_cols) > 0) {
     warning( sprintf('There were %s constant columns removed from input dataset,\n%s', length(const_cols), paste0(const_cols, collapse=', ')) )
     dt = dt[, (const_cols) := NULL]
   }
   return(dt)
 }
+
 # remove date time columns
+#' @import xefun
 check_datetime_cols = function(dt) {
   setDT(dt)
   # datatime columns
-  datetime_cols = names(which(dt[,sapply(.SD, isdatetime)]))
+  datetime_cols = cols_type(dt, 'datetime')
   if (length(datetime_cols) > 0) {
     warning( sprintf('There were %s date/time columns removed from input dataset,\n%s', length(datetime_cols), paste0(datetime_cols, collapse=', ')) )
     dt = dt[,(datetime_cols) := NULL]
   }
   return(dt)
 }
+
 # check categorical columns' unique values
 check_cateCols_uniqueValues = function(dt, var_skip = NULL) {
   setDT(dt)
   # categorical columns
-  cate_cols = names(which(dt[, sapply(.SD, function(x) is.character(x) | is.factor(x))]))
+  cate_cols = cols_type(dt, c('character', 'factor'))
   cate_cols = setdiff(cate_cols, var_skip)
 
   if (length(cate_cols) > 0) {
@@ -59,9 +49,7 @@ rep_blank_na = function(dt) {
   dt = setDT(dt)
 
   # replace black values
-  char_cols = names(which(dt[
-    , sapply(.SD, function(x) is.character(x) | is.factor(x))
-  ]))
+  char_cols = cols_type(dt, c('character', 'factor'))
   if (length(char_cols)>0) {
     # columns have blank value
     blank_cols = names(which(dt[, sapply(.SD, function(x) any(stri_isempty(stri_trim_left(x)))), .SD = char_cols])) # grep('^\\s*$', x)
@@ -251,6 +239,7 @@ check_special_values = function(special_values, xs) {
   return(special_values)
 }
 
+
 # second to hh:mm:ss
 sec_to_hms = function(sec) {
   h = sec %/% 3600
@@ -259,6 +248,7 @@ sec_to_hms = function(sec) {
 
   return(sprintf("%02.f:%02.f:%02.f",h,m,s))
 }
+
 
 # check check_stop_limit
 check_stop_limit = function(stop_limit, xs) {
@@ -282,6 +272,7 @@ check_stop_limit = function(stop_limit, xs) {
   return(sl)
 }
 
+
 # check number of cpu cores
 check_no_cores = function(no_cores, xs_len=10) {
   all_cores = detectCores(logical=F)-1
@@ -301,23 +292,12 @@ check_no_cores = function(no_cores, xs_len=10) {
   return(no_cores)
 }
 
-# y to good bad
-# groupby or dcast
-# groupby is faster via data.table package
-# y_to_goodbad = function(dt, y) {
-#   # dt = data.table(x = rnorm(1e+8), y = sample(c(rep(0,9),1), 1e+8, replace = TRUE))
-#   #
-#   # system.time(
-#   #   dcast(dt, x~y, fun=length, value.var = 'y')
-#   # )
-#   #
-#   # system.time(
-#   #   dt[, .(good=sum(y==0), bad=sum(y==1)), by=x]
-#   # )
-#
-#
-#
-# }
+
+# check woebin or scorecard object
+check_bincard = function(x) {
+  if (inherits(x, 'list') && all(sapply(x, is.data.frame))) x = rbindlist(x, fill = TRUE)
+  setDT(x)
+}
 
 
 menu2 = function(choices, title, chk_rng = TRUE) {
@@ -340,111 +320,3 @@ menu2 = function(choices, title, chk_rng = TRUE) {
 
 
 
-brk_txt2vector = function(brk) {
-  if (is.null(brk) || is.na(brk)) return(brk)
-
-  # v = strsplit(brk, "(?<!\\%),(?!\\%)", perl = T)[[1]]
-  # v = trimws(v)
-  # v = sub('[\'\"](.+)[\'\"]', '\\1', v)
-
-  v = eval(parse(text = sprintf('c(%s)', brk)))
-
-  # dtbrk = data.table(
-  #   brk = brk
-  # )[, strsplit(brk, '%,%')
-  # ][, id := .I
-  # ][, strsplit(V1, ','), by =id
-  # ][, V1 := trimws(V1)
-  # ][, id2 := .I
-  # ][, id3 := id - shift(id,type='lag')]
-  # # dtbrk[1, id3 := dtbrk[2, id3]]
-  #
-  # rowid = which(dtbrk[, (id - shift(id,type = 'lag')) == 1])
-  #
-  # v = rbind(
-  #   data.table(i = rowid, v = paste(dtbrk[rowid-1,V1], dtbrk[rowid,V1], sep = '%,%')),
-  #   dtbrk[-c(rowid, rowid-1), .(i = id2, v=V1)]
-  # )[order(i)][,sub('[\'\"](.+)[\'\"]', '\\1', v)]
-
-  return(v)
-}
-brk_numx_init = function(brk, xvalue, bin_close_right) {
-  brk = as.numeric(brk)
-  brk = setdiff(unique(brk), c(NA, Inf, -Inf))
-  xval = unique(xvalue)
-
-  if (isTRUE(bin_close_right)) {
-    brk=sort(brk[(brk< max(xval, na.rm=TRUE)) & (brk>=min(xval, na.rm=TRUE))])
-  } else {
-    brk=sort(brk[(brk<=max(xval, na.rm=TRUE)) & (brk> min(xval, na.rm=TRUE))])
-  }
-
-  brk = unique(c(-Inf, brk, Inf))
-  return(brk)
-}
-
-
-getarg = function(arg, ...) {
-  x = paste0('scorecard.', arg)
-  kwargs = list(...)
-
-  arg_val = kwargs[[arg]]
-  if (is.null(arg_val)) arg_val = getOption(x)
-
-  return(arg_val)
-}
-
-binpattern = function(type, bin_close_right = NULL) {
-  bin_close_right = getarg('bin_close_right', bin_close_right=bin_close_right)
-
-  pstr_left = list(
-    isbin = '\\[',
-    left_brkp = "^\\[(.*),.+",
-    leftright_brkp = "^\\[(.*), *(.*)\\)",
-    leftrightbrkp_missing = "^\\[(.*), *(.*)\\)((%,%missing)*)",
-    multibin = "^(\\[.+?,).+,(.+?\\))$"
-  )
-
-  pstr_right = list(
-    isbin = '\\(',
-    left_brkp = "^\\((.*),.+",
-    leftright_brkp = "^\\((.*), *(.*)\\]",
-    leftrightbrkp_missing = "^\\((.*), *(.*)\\]((%,%missing)*)",
-    multibin = "^(\\(.+?,).+,(.+?\\])$"
-  )
-
-  if (bin_close_right) {
-    pstr_right[[type]]
-  } else {
-    pstr_left[[type]]
-  }
-}
-
-
-get_brkp_bin = function(bin, bin_close_right=NULL) {
-  bin_close_right = getarg('bin_close_right', bin_close_right=bin_close_right)
-
-  x = '\\1'
-  if (bin_close_right) x = '\\2'
-  as.numeric(sub(binpattern('leftright_brkp',bin_close_right), x, bin))
-}
-
-# bin close right
-get_bcr_bin = function(bins) {
-  . = bcr = bin = NULL
-  if (inherits(bins, 'list')) bins = rbindlist(bins, fill=TRUE)
-  bins[,.(bin)
-  ][(grepl('\\[', bin) & grepl('\\)', bin)) | (grepl('\\(', bin) & grepl('\\]', bin))
-  ][][grepl('\\[', bin), bcr := !grepl('^\\[', bin)
-  ][][!is.na(bcr)
-  ][, unique(bcr)]
-}
-check_bcr = function(bins) {
-  bcr = get_bcr_bin(bins)
-  bin_close_right = getarg('bin_close_right')
-  if (length(bcr) == 1 && (bcr %in% c(TRUE, FALSE)) && bin_close_right != bcr) {
-    bin_close_right = bcr
-    warning(sprintf("The option bin_close_right was set to %s. Resetting by options(scorecard.bin_close_right = %s).\n", bcr, bcr))
-  }
-  return(bin_close_right)
-}